+++ src/classes/Appearance.ts
@@ -2439,7 +2439,7 @@ export class Appearance extends Utils {
                 if (horseCocks == 2) descript += `, ${Appearance.cockNoun(CockTypesEnum.HORSE)}s`;
                 if (dogCocks == 2) descript += `, ${Appearance.cockNoun(CockTypesEnum.DOG)}s`;
                 // Failsafe
-                if (creature.cocks[0].cockType.Index > 2)
+                if (creature.cocks[0].cockType > 2)
                     descript += `, ${Appearance.cockNoun(creature.cocks[0].cockType)}s`;
             }
             // Nonidentical
@@ -2469,7 +2469,7 @@ export class Appearance extends Utils {
                 if (horseCocks == 3) descript += `, ${Appearance.cockNoun(CockTypesEnum.HORSE)}s`;
                 if (dogCocks == 3) descript += `, ${Appearance.cockNoun(CockTypesEnum.DOG)}s`;
                 // Tentacles
-                if (creature.cocks[0].cockType.Index > 2)
+                if (creature.cocks[0].cockType > 2)
                     descript += `, ${Appearance.cockNoun(creature.cocks[0].cockType)}s`;
             } else {
                 descript += Appearance.randomChoice("three ", "group of ");
@@ -2509,7 +2509,7 @@ export class Appearance extends Utils {
                     descript += `${Appearance.cockNoun(CockTypesEnum.HORSE)}s`;
                     descripted = true;
                 }
-                if (creature.cocks[0].cockType.Index > 2) {
+                if (creature.cocks[0].cockType > 2) {
                     descript += `${creature.cockAdjective()}, `;
                     descript += `${Appearance.cockNoun(creature.cocks[0].cockType)}s`;
                     descripted = true;
@@ -2599,7 +2599,7 @@ export class Appearance extends Utils {
                 if (horseCocks == 2) descript += `, ${Appearance.cockNoun(CockTypesEnum.HORSE)}s`;
                 if (dogCocks == 2) descript += `, ${Appearance.cockNoun(CockTypesEnum.DOG)}s`;
                 // Tentacles
-                if (creature.cocks[0].cockType.Index > 2)
+                if (creature.cocks[0].cockType > 2)
                     descript += `, ${Appearance.cockNoun(creature.cocks[0].cockType)}s`;
             }
             // Nonidentical
@@ -2639,7 +2639,7 @@ export class Appearance extends Utils {
                 if (horseCocks == 3) descript += `, ${Appearance.cockNoun(CockTypesEnum.HORSE)}s`;
                 if (dogCocks == 3) descript += `, ${Appearance.cockNoun(CockTypesEnum.DOG)}s`;
                 // Tentacles
-                if (creature.cocks[0].cockType.Index > 2)
+                if (creature.cocks[0].cockType > 2)
                     descript += `, ${Appearance.cockNoun(creature.cocks[0].cockType)}s`; // Not sure what's going on here, referencing index *may* be a bug.
             } else {
                 descript += Appearance.randomChoice("three ", "a group of ");
@@ -2700,7 +2700,7 @@ export class Appearance extends Utils {
                     descripted = true;
                 }
                 // TODO More group cock type descriptions!
-                if (creature.cocks[0].cockType.Index > 2) {
+                if (creature.cocks[0].cockType > 2) {
                     descript += `${Appearance.cockAdjectives(
                         averageLength,
                         averageThickness,