+++ src/classes/Scenes/Areas/Plains/Gnoll.ts
@@ -38,19 +38,19 @@ export class Gnoll extends Monster {
             Math.floor(Math.random() * ((this.player.spe - this.spe) / 4 + 80)) > 80
         ) {
             if (this.player.spe - this.spe < 8)
-                this.outx(`You narrowly avoid ${this.a}[name]'s ${this.weaponVerb}!\n`, false);
+                this.outx(`You narrowly avoid ${this.a}${this.short}'s ${this.weaponVerb}!\n`, false);
             else if (this.player.spe - this.spe >= 8 && this.player.spe - this.spe < 20)
                 this.outx(
-                    `You dodge ${this.a}[name]'s ${this.weaponVerb} with superior quickness!\n`,
+                    `You dodge ${this.a}${this.short}'s ${this.weaponVerb} with superior quickness!\n`,
                     false,
                 );
             else if (this.player.spe - this.spe >= 20)
-                this.outx(`You deftly avoid ${this.a}[name]'s slow ${this.weaponVerb}.\n`, false);
+                this.outx(`You deftly avoid ${this.a}${this.short}'s slow ${this.weaponVerb}.\n`, false);
         }
         // Determine if evaded
         else if (this.player.findPerk(PerkLib.Evade) >= 0 && Gnoll.rand(100) < 10) {
             this.outx(
-                `Using your skills at evading attacks, you anticipate and sidestep ${this.a}[name]'s attack.\n`,
+                `Using your skills at evading attacks, you anticipate and sidestep ${this.a}${this.short}'s attack.\n`,
                 false,
             );
         }
@@ -61,14 +61,14 @@ export class Gnoll extends Monster {
             this.player.armorName == "red, high-society bodysuit"
         ) {
             this.outx(
-                `Using Raphael's teachings, you anticipate and sidestep ${this.a}[name]' attacks.\n`,
+                `Using Raphael's teachings, you anticipate and sidestep ${this.a}${this.short}' attacks.\n`,
                 false,
             );
         }
         // Determine if cat'ed
         else if (this.player.findPerk(PerkLib.Flexibility) >= 0 && Gnoll.rand(100) < 6) {
             this.outx(
-                `With your incredible flexibility, you squeeze out of the way of ${this.a}[name]`,
+                `With your incredible flexibility, you squeeze out of the way of ${this.a}${this.short}`,
             );
             if (this.plural) this.outx("' attacks.\n");
             else this.outx("'s attack.\n");
@@ -90,11 +90,11 @@ export class Gnoll extends Monster {
                 // Due to toughness or amor...
                 else if (Gnoll.rand(this.player.armorDef + this.player.tou) < this.player.armorDef)
                     this.outx(
-                        `Your [armor] absorb and deflect every ${this.weaponVerb} from ${this.a}[name].`,
+                        `Your [armor] absorb and deflect every ${this.weaponVerb} from ${this.a}${this.short}.`,
                     );
                 else
                     this.outx(
-                        `You deflect and block every ${this.weaponVerb} ${this.a}[name] throws at you.`,
+                        `You deflect and block every ${this.weaponVerb} ${this.a}${this.short} throws at you.`,
                     );
             }
             // everyone else
@@ -218,19 +218,19 @@ export class Gnoll extends Monster {
             Math.floor(Math.random() * ((this.player.spe - this.spe) / 4 + 80)) > 80
         ) {
             if (this.player.spe - this.spe < 8)
-                this.outx(`You narrowly avoid ${this.a}[name]'s ${this.weaponVerb}!\n`, false);
+                this.outx(`You narrowly avoid ${this.a}${this.short}'s ${this.weaponVerb}!\n`, false);
             else if (this.player.spe - this.spe >= 8 && this.player.spe - this.spe < 20)
                 this.outx(
-                    `You dodge ${this.a}[name]'s ${this.weaponVerb} with superior quickness!\n`,
+                    `You dodge ${this.a}${this.short}'s ${this.weaponVerb} with superior quickness!\n`,
                     false,
                 );
             else if (this.player.spe - this.spe >= 20)
-                this.outx(`You deftly avoid ${this.a}[name]'s slow ${this.weaponVerb}.\n`, false);
+                this.outx(`You deftly avoid ${this.a}${this.short}'s slow ${this.weaponVerb}.\n`, false);
         }
         // Determine if evaded
         else if (this.player.findPerk(PerkLib.Evade) >= 0 && Gnoll.rand(100) < 10) {
             this.outx(
-                `Using your skills at evading attacks, you anticipate and sidestep ${this.a}[name]'s attack.\n`,
+                `Using your skills at evading attacks, you anticipate and sidestep ${this.a}${this.short}'s attack.\n`,
                 false,
             );
         }
@@ -241,14 +241,14 @@ export class Gnoll extends Monster {
             this.player.armorName == "red, high-society bodysuit"
         ) {
             this.outx(
-                `Using Raphael's teachings, you anticipate and sidestep ${this.a}[name]' attacks.\n`,
+                `Using Raphael's teachings, you anticipate and sidestep ${this.a}${this.short}' attacks.\n`,
                 false,
             );
         }
         // Determine if cat'ed
         else if (this.player.findPerk(PerkLib.Flexibility) >= 0 && Gnoll.rand(100) < 6) {
             this.outx(
-                `With your incredible flexibility, you squeeze out of the way of ${this.a}[name]`,
+                `With your incredible flexibility, you squeeze out of the way of ${this.a}${this.short}`,
             );
             if (this.plural) this.outx("' attacks.\n");
             else this.outx("'s attack.\n");
@@ -270,11 +270,11 @@ export class Gnoll extends Monster {
                 // Due to toughness or amor...
                 else if (Gnoll.rand(this.player.armorDef + this.player.tou) < this.player.armorDef)
                     this.outx(
-                        `Your [armor] absorb and deflect every ${this.weaponVerb} from ${this.a}[name].`,
+                        `Your [armor] absorb and deflect every ${this.weaponVerb} from ${this.a}${this.short}.`,
                     );
                 else
                     this.outx(
-                        `You deflect and block every ${this.weaponVerb} ${this.a}[name] throws at you.`,
+                        `You deflect and block every ${this.weaponVerb} ${this.a}${this.short} throws at you.`,
                     );
             }
             // everyone else
@@ -375,7 +375,7 @@ export class Gnoll extends Monster {
             );
             if (this.statusAffectv1(StatusAffects.Constricted) <= 0) {
                 this.outx(
-                    `  ${this.capitalA}[name] proves to be too much for your tail to handle, breaking free of your tightly bound coils.`,
+                    `  ${this.capitalA}${this.short} proves to be too much for your tail to handle, breaking free of your tightly bound coils.`,
                 );
                 this.removeStatusAffect(StatusAffects.Constricted);
             }
@@ -420,22 +420,22 @@ export class Gnoll extends Monster {
                 Math.floor(Math.random() * ((this.player.spe - this.spe) / 4 + 80)) > 80
             ) {
                 if (this.player.spe - this.spe < 8)
-                    this.outx(`You narrowly avoid ${this.a}[name]'s ${this.weaponVerb}!\n`, false);
+                    this.outx(`You narrowly avoid ${this.a}${this.short}'s ${this.weaponVerb}!\n`, false);
                 else if (this.player.spe - this.spe >= 8 && this.player.spe - this.spe < 20)
                     this.outx(
-                        `You dodge ${this.a}[name]'s ${this.weaponVerb} with superior quickness!\n`,
+                        `You dodge ${this.a}${this.short}'s ${this.weaponVerb} with superior quickness!\n`,
                         false,
                     );
                 else if (this.player.spe - this.spe >= 20)
                     this.outx(
-                        `You deftly avoid ${this.a}[name]'s slow ${this.weaponVerb}.\n`,
+                        `You deftly avoid ${this.a}${this.short}'s slow ${this.weaponVerb}.\n`,
                         false,
                     );
             }
             // Determine if evaded
             else if (this.player.findPerk(PerkLib.Evade) >= 0 && Gnoll.rand(100) < 10) {
                 this.outx(
-                    `Using your skills at evading attacks, you anticipate and sidestep ${this.a}[name]'s attack.\n`,
+                    `Using your skills at evading attacks, you anticipate and sidestep ${this.a}${this.short}'s attack.\n`,
                     false,
                 );
             }
@@ -446,14 +446,14 @@ export class Gnoll extends Monster {
                 this.player.armorName == "red, high-society bodysuit"
             ) {
                 this.outx(
-                    `Using Raphael's teachings, you anticipate and sidestep ${this.a}[name]' attacks.\n`,
+                    `Using Raphael's teachings, you anticipate and sidestep ${this.a}${this.short}' attacks.\n`,
                     false,
                 );
             }
             // Determine if cat'ed
             else if (this.player.findPerk(PerkLib.Flexibility) >= 0 && Gnoll.rand(100) < 6) {
                 this.outx(
-                    `With your incredible flexibility, you squeeze out of the way of ${this.a}[name]`,
+                    `With your incredible flexibility, you squeeze out of the way of ${this.a}${this.short}`,
                 );
                 if (this.plural) this.outx("' attacks.\n");
                 else this.outx("'s attack.\n");
@@ -477,11 +477,11 @@ export class Gnoll extends Monster {
                         Gnoll.rand(this.player.armorDef + this.player.tou) < this.player.armorDef
                     )
                         this.outx(
-                            `Your [armor] absorb and deflect every ${this.weaponVerb} from ${this.a}[name].`,
+                            `Your [armor] absorb and deflect every ${this.weaponVerb} from ${this.a}${this.short}.`,
                         );
                     else
                         this.outx(
-                            `You deflect and block every ${this.weaponVerb} ${this.a}[name] throws at you.`,
+                            `You deflect and block every ${this.weaponVerb} ${this.a}${this.short} throws at you.`,
                         );
                 }
                 // everyone else