+++ src/classes/Scenes/Monsters/GoblinScene.ts
@@ -1044,7 +1044,7 @@ export class GoblinScene extends BaseContent {
             else if (this.player.hasKnot(0)) this.outx("stretching painfully around your knot");
             else if (
                 this.player.cocks[0].cockType == CockTypesEnum.HORSE ||
-                this.player.cocks[0].cockType.Index > 3
+                this.player.cocks[0].cockType > 3
             )
                 this.outx("feeling painfully tight around you");
             this.outx(
@@ -1132,7 +1132,7 @@ export class GoblinScene extends BaseContent {
             else if (this.player.hasKnot(0)) this.outx("stretching tightly around your knot");
             else if (
                 this.player.cocks[0].cockType == CockTypesEnum.HORSE ||
-                this.player.cocks[0].cockType.Index > 3
+                this.player.cocks[0].cockType > 3
             )
                 this.outx("feeling absolutely perfect around you");
             this.outx(