+++ src/classes/Scenes/NPCs/Amily.ts
@@ -101,11 +101,11 @@ export class Amily extends Monster {
             // Due to toughness or amor...
             if (Amily.rand(this.player.armorDef + this.player.tou) < this.player.armorDef)
                 this.outx(
-                    `Your [armor] absorb and deflect every ${this.weaponVerb} from ${this.a}[name].`,
+                    `Your [armor] absorb and deflect every ${this.weaponVerb} from ${this.a}${this.short}.`,
                 );
             else
                 this.outx(
-                    `You deflect and block every ${this.weaponVerb} ${this.a}[name] throws at you.`,
+                    `You deflect and block every ${this.weaponVerb} ${this.a}${this.short} throws at you.`,
                 );
         }
         // Got hit!
@@ -117,12 +117,12 @@ export class Amily extends Monster {
             if (this.lustVuln > 0 && this.player.armorName == "barely-decent bondage straps") {
                 if (!this.plural)
                     this.outx(
-                        `\n${this.capitalA}[name] brushes against your exposed skin and jerks back in surprise, coloring slightly from seeing so much of you revealed.`,
+                        `\n${this.capitalA}${this.short} brushes against your exposed skin and jerks back in surprise, coloring slightly from seeing so much of you revealed.`,
                         false,
                     );
                 else
                     this.outx(
-                        `\n${this.capitalA}[name] brush against your exposed skin and jerk back in surprise, coloring slightly from seeing so much of you revealed.`,
+                        `\n${this.capitalA}${this.short} brush against your exposed skin and jerk back in surprise, coloring slightly from seeing so much of you revealed.`,
                         false,
                     );
                 this.lust += 10 * this.lustVuln;